% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BPLR.R
\name{BPLR}
\alias{BPLR}
\title{Estimating the ROC curve}
\usage{
BPLR(x, y, method = "Bernstein", nss = 10^4, type = 3)
}
\arguments{
\item{x}{vector contains the sample of biomarkers from the "healthy" group}

\item{y}{vector contains the sample of biomarkers from the "diseased" group}

\item{method}{the method to be used to estimate ROC curve and its summary statistics.
The default method "Bernstein" uses the Bernstein polynomials to approximate the log of density ratio incorporating the
likelihood ratio ordering.}

\item{nss}{control the number of point estimates of the ROC curve in the range [0,1], default is 10^4.}

\item{type}{control the type of Bernstein polynomials used to approximate the log density ratio.
"type=1" uses the Bernstein polynomial of x and y. "type=2" uses the Bernstein polynomial of log(x) and log(y).
"type=3" uses both Bernstein polynomial of x, y and log(x), log(y).
"type=4" automatically choose between type=1 and 2.
 In each case, the order of the Bernstein polynomials being used is selected by a BIC criteria.}
}
\description{
To estimate ROC curve and its summary statistics including AUC, Youden index and cutoff point based on a single biomarker.
Default Bernstein polynomial approach deals with the case that the density ratio of x and y is monotone.
Other approaches are also available. See more details below.
}
\details{
The function
}
\examples{
x=rnorm(100,10,1)
y=rnrom(100,12,1)
BPLR(x,y,method="Bernstein",nss=10^4,type=1)

}
